operator practice

+++ OperatorPratice/src/operatorpratice/OperatorPratice.java
@@ -16,14 +16,14 @@ public class OperatorPratice {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-    double testScore1 = 90;
-    double testScore2 = 90;
-    double testScore3 = 89;
+    int testScore1 = 90;
+    int testScore2 = 90;
+    int testScore3 = 89;
     System.out.println("The score for test 1 is " + testScore1 + ".");
     System.out.println("The score for test 2 is " + testScore2 + ".");
     System.out.println("The score for test 3 is " + testScore3 + ".");
     double average;
-    average = (testScore1 + testScore2 + testScore3)/3;
+    average = (testScore1 + testScore2 + testScore3)/3.0;
     System.out.println("The average test score is " + average + "." );
     }
